requirements.txt
alembic==1.0.11
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.20.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.14
MarkupSafe==1.1.1
passlib==1.7.1
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
SQLAlchemy-Utils==0.34.1
Werkzeug==0.15.4
